+++ projects/ui/src/Components/Apps/PageContent/AppSummaryCards/AppSummaryListCard.tsx
@@ -1,7 +1,7 @@
 import { Box } from "@mantine/core";
 import { NavLink } from "react-router-dom";
 import { Icon } from "../../../../Assets/Icons";
-import { GridCardStyles } from "../../../../Styles/shared/GridCard.style";
+import { CardStyles } from "../../../../Styles/shared/Card.style";
 import { ListCardStyles } from "../../../../Styles/shared/ListCard.style";
 import { getAppDetailsLink } from "../../../../Utility/link-builders";
 import { AppWithTeam } from "../AppsList";
@@ -18,19 +18,9 @@ export function AppSummaryListCard({ app }: { app: AppWithTeam }) {
             <Icon.WrenchGear className="colorIt" />
           </div>
           <Box p={"30px"}>
-            <GridCardStyles.Title>{app.name}</GridCardStyles.Title>
-            <GridCardStyles.Description>
-              {app.description}
-            </GridCardStyles.Description>
+            <CardStyles.Title>{app.name}</CardStyles.Title>
+            <CardStyles.Description>{app.description}</CardStyles.Description>
           </Box>
-          {/* <div className="details">
-            <Flex direction={"column"}>
-              <GridCardStyles.Title>{app.name}</GridCardStyles.Title>
-              <GridCardStyles.Description>
-                {app.description}
-              </GridCardStyles.Description>
-            </Flex>
-          </div> */}
         </div>
         <div className="footer">
           <div className="metaInfo">
